chore: update demo script

+++ demo.sh
@@ -97,8 +97,8 @@ if [ -z "$NUM_PROCESSING_REQUESTS" ]; then
   NUM_PROCESSING_REQUESTS=$DEFAULT_NUM_PROCESSING_REQUESTS
 fi
 
-if ! echo "$NUM_WORKERS" | grep -qE '^[0-9]+$' || [ "$NUM_WORKERS" -lt 1 ]; then
-  exit_on_error "Number of workers must be a positive integer"
+if ! echo "$NUM_WORKERS" | grep -qE '^[0-9]+$' || [ "$NUM_WORKERS" -lt 0 ]; then
+  exit_on_error "Number of workers must be a non-negative integer"
 fi
 
 if ! echo "$NUM_PROCESSING_REQUESTS" | grep -qE '^[0-9]+$' || [ "$NUM_PROCESSING_REQUESTS" -lt 1 ]; then
@@ -398,6 +398,11 @@ check_workers_available() {
   local WORKERS_RESPONSE=$(call_backend "GET" "/admin/worker")
   local WORKERS_WITH_LAST_SEEN=$(echo "$WORKERS_RESPONSE" | jq -r '.edges[] | select(.node.last_seen_at != null) | .node.id' | wc -l)
   
+  if [ "$WORKERS_WITH_LAST_SEEN" -eq 0 ]; then
+    echo "[INFO] No workers detected yet..."
+    return 1
+  fi
+  
   if [ "$WORKERS_WITH_LAST_SEEN" -lt "$NUM_WORKERS" ]; then
     echo "[INFO] Waiting for workers to be active... (Active workers: $WORKERS_WITH_LAST_SEEN/$NUM_WORKERS)"
     return 1
@@ -525,25 +530,54 @@ if [ -z "$WORKER_REGISTRATION_TOKEN" ] || [ "$WORKER_REGISTRATION_TOKEN" = "null
   exit_on_error "Failed to create worker registration token."
 fi
 echo "Worker Registration Token: $WORKER_REGISTRATION_TOKEN"
+echo
+echo "[INFO] To start a worker on a remote server, run this command:"
+echo "docker run --rm --network host \\"
+echo "  -v /var/run/docker.sock:/var/run/docker.sock \\"
+echo "  -v autodroid_worker_data:/usr/app/temp:rw \\"
+echo "  --pull always $WORKER_IMAGE_NAME \\"
+echo "  -u \"http://<<THIS_BACKEND_MACHINE_NETWORK_IP>>:$PORT\" \\"
+echo "  -n \"remote_worker_X\" \\"
+echo "  -t \"$WORKER_REGISTRATION_TOKEN\""
+echo
+echo "Remember to replace <<THIS_BACKEND_MACHINE_NETWORK_IP>> with the IP address of the target machine in the network."
+echo "Replace X with the worker number, for example, if you want to start 2 workers, you should run the command 2 times, with remote_worker_1 and remote_worker_2 as the worker names."
+echo "Also create the folder /usr/app/temp in the target machine before running the command."
+echo
+echo "[INFO] The script will start $NUM_WORKERS local worker(s)."
+echo "[INFO] Press Enter to continue after starting all workers (local and remote)..."
+read dummy
 
 
 #
 # STEP 3
 #
 step "Step 3" "Start $NUM_WORKERS worker container(s)."
-for i in $(seq 1 "$NUM_WORKERS"); do
-  CONTAINER_NAME="${CONTAINER_NAME_PREFIX}_${i}"
-  
-  docker run --name "$CONTAINER_NAME" --rm --network host \
-    -v /var/run/docker.sock:/var/run/docker.sock \
-    -v "$VOLUME_NAME":/usr/app/temp:rw \
-    --pull always "$WORKER_IMAGE_NAME" \
-    -u "http://host.docker.internal:$PORT" \
-    -n "$CONTAINER_NAME" \
-    -t "$WORKER_REGISTRATION_TOKEN" &
-  
-  echo "Worker container $CONTAINER_NAME started."
-done
+if [ "$NUM_WORKERS" -gt 0 ]; then
+  for i in $(seq 1 "$NUM_WORKERS"); do
+    CONTAINER_NAME="${CONTAINER_NAME_PREFIX}_${i}"
+    
+    docker run --name "$CONTAINER_NAME" --rm --network host \
+      -v /var/run/docker.sock:/var/run/docker.sock \
+      -v "$VOLUME_NAME":/usr/app/temp:rw \
+      --pull always "$WORKER_IMAGE_NAME" \
+      -u "http://host.docker.internal:$PORT" \
+      -n "$CONTAINER_NAME" \
+      -t "$WORKER_REGISTRATION_TOKEN" &
+    
+    echo "Worker container $CONTAINER_NAME started."
+  done
+
+  while ! check_workers_available; do
+    sleep 5
+  done
+else
+  echo "[INFO] No local workers will be started."
+  echo "[INFO] Waiting for at least one remote worker to be available..."
+  while ! check_workers_available; do
+    sleep 5
+  done
+fi
 
 #
 # STEP 4
@@ -599,10 +633,6 @@ fi
 # STEP 6
 #
 step "Step 6" "Request processing for the dataset(s)."
-while ! check_workers_available; do
-  sleep 5
-done
-
 PROCESSING_REQUEST_IDS=""
 
 for i in $(seq 1 "$NUM_PROCESSING_REQUESTS"); do
